Rename Project to 'Renamer' and add gradle plugin
