+++ .github/workflows/plugin-publish.yml
@@ -1,0 +1,31 @@
+name: Publish Gradle Plugin
+
+on:
+  push:
+    branches: [ 'master' ]
+    paths:
+      - 'renamer-gradle/**'
+      - '!.github/workflows/**'
+      - '!docs/**'
+      - '!README.md'
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    uses: MinecraftForge/SharedActions/.github/workflows/gradle.yml@v0
+    with:
+      java: 17
+      gradle_tasks: ':renamer-gradle:check :renamer-gradle-demo:assemble :renamer-gradle:publish :renamer-gradle:publishPlugins'
+      project_path: 'renamer-gradle'
+    secrets:
+      DISCORD_WEBHOOK: ${{ secrets.DISCORD_WEBHOOK }}
+      MAVEN_USER: ${{ secrets.MAVEN_USER }}
+      MAVEN_PASSWORD: ${{ secrets.MAVEN_PASSWORD }}
+      PROMOTE_ARTIFACT_WEBHOOK: ${{ secrets.PROMOTE_ARTIFACT_WEBHOOK }}
+      PROMOTE_ARTIFACT_USERNAME: ${{ secrets.PROMOTE_ARTIFACT_USERNAME }}
+      PROMOTE_ARTIFACT_PASSWORD: ${{ secrets.PROMOTE_ARTIFACT_PASSWORD }}
+      GRADLE_PUBLISH_KEY: ${{ secrets.GRADLE_PUBLISH_KEY }}
+      GRADLE_PUBLISH_SECRET: ${{ secrets.GRADLE_PUBLISH_SECRET }}
+      GRADLE_CACHE_KEY: ${{ secrets.GRADLE_CACHE_KEY }}
