+++ .github/workflows/publish.yml
@@ -3,6 +3,12 @@ name: Publish
 on:
   push:
     branches: [ "master" ]
+    paths-ignore:
+      - 'renamer-gradle/**'
+      - 'renamer-gradle-demo/**'
+      - '.github/workflows/**'
+      - 'docs/**'
+      - 'README.md'
 
 permissions:
   contents: read
@@ -12,8 +18,8 @@ jobs:
     uses: MinecraftForge/SharedActions/.github/workflows/gradle.yml@main
     with:
       java: 17
-      gradle_tasks: "publish"
-      artifact_name: "ForgeAutoRenamingTool"
+      gradle_tasks: "check publish"
+      artifact_name: "renamer"
     secrets:
       DISCORD_WEBHOOK: ${{ secrets.DISCORD_WEBHOOK }}
       PROMOTE_ARTIFACT_WEBHOOK: ${{ secrets.PROMOTE_ARTIFACT_WEBHOOK }}
@@ -21,3 +27,4 @@ jobs:
       PROMOTE_ARTIFACT_PASSWORD: ${{ secrets.PROMOTE_ARTIFACT_PASSWORD }}
       MAVEN_USER: ${{ secrets.MAVEN_USER }}
       MAVEN_PASSWORD: ${{ secrets.MAVEN_PASSWORD }}
+      GRADLE_CACHE_KEY: ${{ secrets.GRADLE_CACHE_KEY }}
