+++ LICENSE-header.txt
@@ -1,2 +1,2 @@
-Copyright (c) Forge Development LLC and contributors
+Copyright (c) Forge Development LLC
 SPDX-License-Identifier: LGPL-2.1-only
