+++ renamer-gradle/src/main/java/net/minecraftforge/renamer/gradle/RenameJar.java
@@ -1,3 +1,7 @@
+/*
+ * Copyright (c) Forge Development LLC
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
 package net.minecraftforge.renamer.gradle;
 
 import net.minecraftforge.gradleutils.shared.ToolExecBase;
