+++ renamer-gradle/src/main/java/net/minecraftforge/renamer/gradle/Tools.java
@@ -9,5 +9,5 @@ import net.minecraftforge.gradleutils.shared.Tool;
 final class Tools {
     private Tools() { }
 
-    static final Tool RENAMER = Tool.ofForge("renamer", "net.minecraftforge:ForgeAutoRenamingTool:1.1.2:all", 8);
+    static final Tool RENAMER = Tool.ofForge("classes", "net.minecraftforge:renamer:2.0.0:all", 8);
 }
