+++ settings.gradle
@@ -2,8 +2,7 @@ plugins {
     id 'org.gradle.toolchains.foojay-resolver-convention' version '1.0.0'
 }
 
-// TODO [Renamer] Change to 'renamer'
-rootProject.name = 'ForgeAutoRenamingTool'
+rootProject.name = 'renamer'
 
 includeBuild 'renamer-gradle'
 includeBuild 'renamer-gradle-demo'
