+++ src/main/java/net/minecraftforge/renamer/api/Renamer.java
@@ -1,14 +1,14 @@
 /*
- * Copyright (c) Forge Development LLC and contributors
+ * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.fart.api;
+package net.minecraftforge.renamer.api;
 
 import java.io.Closeable;
 import java.io.File;
 import java.util.function.Consumer;
 
-import net.minecraftforge.fart.internal.RenamerBuilder;
+import net.minecraftforge.renamer.internal.RenamerBuilder;
 
 /**
  * A {@code Renamer} is used to run generic transformers on a JAR file.
