+++ src/main/java/net/minecraftforge/renamer/api/SourceFixerConfig.java
@@ -1,8 +1,8 @@
 /*
- * Copyright (c) Forge Development LLC and contributors
+ * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.fart.api;
+package net.minecraftforge.renamer.api;
 
 /**
  * Source file naming strategy for {@link Transformer#sourceFixerFactory(SourceFixerConfig)}.
