+++ src/main/java/net/minecraftforge/renamer/internal/ClassLoaderClassProvider.java
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) Forge Development LLC and contributors
+ * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.fart.internal;
+package net.minecraftforge.renamer.internal;
 
-import net.minecraftforge.fart.api.ClassProvider;
 import org.jetbrains.annotations.Nullable;
+
+import net.minecraftforge.renamer.api.ClassProvider;
 
 import java.io.IOException;
 import java.util.Optional;
