+++ src/main/java/net/minecraftforge/renamer/internal/EntryImpl.java
@@ -1,10 +1,10 @@
 /*
- * Copyright (c) Forge Development LLC and contributors
+ * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.fart.internal;
+package net.minecraftforge.renamer.internal;
 
-import net.minecraftforge.fart.api.Transformer;
+import net.minecraftforge.renamer.api.Transformer;
 
 public abstract class EntryImpl implements Transformer.Entry {
     private final String name;
