+++ src/main/java/net/minecraftforge/renamer/internal/SourceFixer.java
@@ -1,11 +1,12 @@
 /*
- * Copyright (c) Forge Development LLC and contributors
+ * Copyright (c) Forge Development LLC
  * SPDX-License-Identifier: LGPL-2.1-only
  */
-package net.minecraftforge.fart.internal;
+package net.minecraftforge.renamer.internal;
 
-import net.minecraftforge.fart.api.SourceFixerConfig;
 import org.objectweb.asm.ClassVisitor;
+
+import net.minecraftforge.renamer.api.SourceFixerConfig;
 
 public final class SourceFixer extends OptionalChangeTransformer {
